feat: adiciona módulo completo de importação de arquivos e integra criação automática de pessoas e contas
- Implementa fluxo de upload (.csv/.xlsx) com mapeamento para ImporterDTO
- Adiciona FileImporter, fábrica de importadores e parsers de datas/números/cpf-cnpj
- Integra importação com criação automática de Pessoa (PF/PJ) e Conta
- Atualiza DTOs (id Integer → Long) e adiciona formaPagamento em Conta
- Ajusta mappers (Usuario e Conta) para suporte ao novo fluxo
- Refatora controllers e reorganiza pacote de documentação
- Atualiza serviços e repositórios para uso de Long
- Configura Hibernate ddl-auto para create

BREAKING CHANGE: mudança de tipos de ID, ajustes estruturais em DTOs, mappers e serviços; recriação do banco necessária.

+++ src/main/java/br/ifrn/edu/ProfitFlow/controller/AuthController.java
@@ -1,7 +1,7 @@
 package br.ifrn.edu.ProfitFlow.controller;
 
 
-import br.ifrn.edu.ProfitFlow.config.docs.AuthControllerDocs;
+import br.ifrn.edu.ProfitFlow.controller.docs.AuthControllerDocs;
 import br.ifrn.edu.ProfitFlow.dto.request.RequestPessoaDTO;
 import br.ifrn.edu.ProfitFlow.dto.response.ResponsePessoaDTO;
 import br.ifrn.edu.ProfitFlow.models.Usuario;
@@ -52,7 +52,7 @@ public class AuthController implements AuthControllerDocs {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Optional<Usuario>> findById(@PathVariable Integer id){
+    public ResponseEntity<Optional<Usuario>> findById(@PathVariable Long id){
         return ResponseEntity.ok(usuarioService.findById(id));
     }
