+++ src/main/java/br/ifrn/edu/ProfitFlow/repository/ContaRepository.java
@@ -3,5 +3,5 @@ package br.ifrn.edu.ProfitFlow.repository;
 import br.ifrn.edu.ProfitFlow.models.Conta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContaRepository extends JpaRepository<Conta, Integer> {
+public interface ContaRepository extends JpaRepository<Conta, Long> {
 }
